src/requirements.txt
Django==1.10.6
django-webpack-loader==0.4.1
djangorestframework==3.6.2
